+++ README.md
@@ -33,3 +33,13 @@ The forms are [Data Driven Forms](https://data-driven-forms.org/), configured th
     - [ ] P1. Keyword search through applications (i.e. to find ones related to health, finance, climate)
     - [ ] The ability to export project applications (same as the ability to review forms individually but also to have in a form you can send people) 
     - [ ] P2 - A review tool for CoP members who can vote yes/no on projects? 
+
+## Configuration
+
+Set the following environment variables in a `.env.local` file:
+```bash
+ACCESS_TOKEN={YOUR_GITHUB_TOKEN_HERE}
+GITHUB_OWNER=lacabra
+GITHUB_REPO=submission-digitalpublicgoods
+GITHUB_BRANCH=main
+```
